% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_occ_plot.R
\name{abun_occ_plot}
\alias{abun_occ_plot}
\title{abun_occ_plot}
\usage{
abun_occ_plot(otu, sample, mini_abun, threshold, height, weight)
}
\arguments{
\item{otu}{a community count data matrix,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{mini_abun}{a value indicating whether the otu is present,}

\item{threshold}{a value indicating the final percent increase in beta-diversity,}

\item{height}{the height of the figure,}

\item{weight}{the width of the figure,}
}
\value{
A scatter plot showing the relationship between occupancy and abundance with color denoting and core or non-core otu.
}
\description{
Plots occupancy vs abunbance.
}
\details{
Plots occupancy vs abunbance.
}
\examples{
 abun_occ_plot(otu, sample, mini_abun=0, threshold=0.02, height=400, weight=400)
}
