% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_plot.R
\name{abundance_plot}
\alias{abundance_plot}
\title{abundance_plot}
\usage{
abundance_plot(otu, taxa, sample, sample_group, rank)
}
\arguments{
\item{otu}{a community count data matrix,}

\item{taxa}{a taxa information matrix. The rownames must match the OTU names (taxa_names) of the otu,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{sample_group}{the name of a column that containing the sample type/group,}

\item{rank}{a taxonomic rank (Phylum or Class or Order or Family or Genus or Species),}
}
\value{
two subfigures containing abundance among samples and groups (sample_group) at specific taxonomic level.
}
\description{
Plots abundance among samples and groups (sample_group) at specific taxonomic level.
}
\details{
Plots abundance among samples and groups (sample_group) at specific taxonomic level.
}
\examples{
 abundance_plot(otu, taxa, sample, sample_group, 'Genus')
}
