% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_rank_plot.R
\name{bc_rank_plot}
\alias{bc_rank_plot}
\title{bc_rank_plot}
\usage{
bc_rank_plot(otu, sample, mini_abun, threshold, sample_name, sample_group)
}
\arguments{
\item{otu}{a community count data matrix,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{mini_abun}{a value indicating whether the otu is present,}

\item{threshold}{a value indicating the final percent increase in beta-diversity,}

\item{sample_name}{the name of a column that containing the sample name,}

\item{sample_group}{the name of a column that containing the sample type/group,}
}
\value{
A scatter plot showing the relationship between BC similarity and ranked otu with color denoting and core or non-core otu and a line indicating the threshold.
}
\description{
Plots BC similarity vs ranked otu.
}
\details{
Plots BC similarity vs ranked otu.
}
\examples{
 bc_rank_plot(otu, sample, mini_abun=0, threshold=0.02, sample_name, sample_group)
}
