% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_func_compare.R
\name{count_func_compare}
\alias{count_func_compare}
\title{count_func_compare}
\usage{
count_func_compare(otu, funct, taxa, rank)
}
\arguments{
\item{otu}{a community count data matrix with samples in rows and OTUs/taxa in column,}

\item{taxa}{a taxa information matrix. The rownames must match the OTU names (taxa_names) of the otu,}

\item{rank}{a taxonomic rank (Phylum or Class or Order or Family or Genus or Species),}

\item{functional_profile}{a community count data matrix with samples in rows and OTUs/taxa in column,}
}
\value{
Arrow plot showing the procruste rotation comparing otu and functions and the overlap between otu and functions representations using spls.
}
\description{
Plots arrow plots showing the procruste rotation comparing otu and functions and the overlap between otu and functions representations using spls.
}
\details{
Plots arrow plots showing the procruste rotation comparing otu and functions and the overlap between otu and functions representations using spls.
}
\examples{
 count_func_compare(otu, funct, taxa, rank='Class')
}
