% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_plot.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{heatmap_plot}
\usage{
heatmap_plot(otu, taxa, sample, rank)
}
\arguments{
\item{otu}{a community count data matrix,}

\item{taxa}{a taxa information matrix. The rownames must match the OTU names (taxa_names) of the otu,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{rank}{a taxonomic rank (Phylum or Class or Order or Family or Genus or Species),}
}
\value{
four heatmap with clustering abundance based on different transformation methods.
}
\description{
Plots heatmap based on count data, relative abundance data, clr transformation data or log10 transformation data.
}
\details{
Plots abundance among samples and groups (SampleType) at specific taxonomic level.
}
\examples{
 heatmap_plot(otu, taxa, sample, 'Genus')
}
