% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_func_occ_diff_plot.R
\name{mean_func_occ_diff_plot}
\alias{mean_func_occ_diff_plot}
\title{mean_func_occ_diff_plot}
\usage{
mean_func_occ_diff_plot(
  functional_profile,
  sample,
  min_tpm,
  max_tpm,
  tpm_step,
  min_thre,
  max_thre,
  thre_step
)
}
\arguments{
\item{functional_profile}{a community count data matrix with samples in rows and OTUs/taxa in column,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{max_tpm}{the maximum TPM of the loop,}

\item{max_thre}{the maximum threshold of the loop,}

\item{thre_step}{the threshold step of the loop,}

\item{mini_tpm}{the minimum TPM of the loop,}

\item{abun_step}{the TPM step of the loop,}

\item{mini_thre}{the minimum threshold of the loop,}
}
\value{
A heatmap showing the mean occurrence frequency difference between core and non-core functional otu under different minimal tpm and thresholds.
}
\description{
Plots a heatmap showing the mean occurrence frequency difference between core and non-core functional otu under different minimal tpm and thresholds.
}
\details{
Plots a heatmap showing the mean occurrence frequency difference between core and non-core functional otu under different minimal tpm and thresholds.
}
\examples{
 mean_func_occ_diff_plot(functional_profile, sample, min_tpm = 0, max_tpm = 4000, tpm_step = 100, min_thre = 0, max_thre = 0.1, thre_step = 0.02)
}
