% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_shared_ne_plot.R
\name{networks_shared_ne_plot}
\alias{networks_shared_ne_plot}
\title{networks_shared_ne_plot}
\usage{
networks_shared_ne_plot(
  otu,
  taxa,
  sample,
  pre_threshold,
  fdr_threshold,
  cor_threshold,
  permutation,
  propr = TRUE,
  sparcc = TRUE,
  cclasso = TRUE,
  propr_col,
  sparcc_col,
  cclasso_col
)
}
\arguments{
\item{otu}{a community count data matrix,}

\item{taxa}{a taxa information matrix. The rownames must match the OTU names (taxa_names) of the otu,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{pre_threshold}{occurrence frequency,}

\item{fdr_threshold}{fdr threshold,}

\item{cor_threshold}{association threshold,}

\item{permutation}{permutation number of the association methods,}

\item{propr}{a logical value indicating whether to run propr,}

\item{sparcc}{a logical value indicating whether to run sparcc,}

\item{cclasso}{a logical value indicating whether to run cclasso,}

\item{propr_col}{the color of the properties of propr network,}

\item{sparcc_col}{the color of the properties of sparcc network,}

\item{cclasso_col}{the color of the properties of cclasso network,}
}
\value{
Upset plots showing the common/shared nodes/edges between networks.
}
\description{
Constructs the co-occurrence networks and calculates the common/shared nodes/edges among networks from different methods and visualization.
}
\details{
Constructs the co-occurrence networks and calculates the common/shared nodes/edges among networks from different methods and visualization.
}
\examples{
 networks_shared_ne_plot(otu, taxa, sample, pre_threshold=0.5, fdr_threshold=0.1, cor_threshold=0.6, permutation=100, propr=TRUE, sparcc=TRUE, cclasso=TRUE, propr_col='darkcyan', sparcc_col='firebrick3', cclasso_col='goldenrod3')
}
