% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_taxa_role.R
\name{networks_taxa_role}
\alias{networks_taxa_role}
\title{networks_taxa_role}
\usage{
networks_taxa_role(
  otu,
  taxa,
  sample,
  pre_threshold,
  fdr_threshold,
  cor_threshold,
  permutation,
  cluster_method,
  propr = TRUE,
  sparcc = TRUE,
  cclasso = TRUE
)
}
\arguments{
\item{otu}{a community count data matrix with samples in rows and OTUs/taxa in column,}

\item{taxa}{a taxa information matrix. The rownames must match the OTU names (taxa_names) of the otu,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{pre_threshold}{occurrence frequency,}

\item{fdr_threshold}{fdr threshold,}

\item{cor_threshold}{association threshold,}

\item{permutation}{permutation number of the association methods,}

\item{cluster_method}{module clustering methods, including cluster_walktrap, cluster_edge_betweenness, cluster_fast_greedy and cluster_spinglass,}

\item{propr}{a logical value indicating whether to run propr,}

\item{sparcc}{a logical value indicating whether to run sparcc,}

\item{cclasso}{a logical value indicating whether to run cclasso,}
}
\value{
A dataframe containing the role of the taxa in a network.
}
\description{
Determine the role of the taxa in a network.
}
\details{
Determine the role of the taxa in a network.
}
\examples{
 networks_taxa_role(otu, taxa, sample, pre_threshold=0.5, fdr_threshold=0.1, cor_threshold=0.6, permutation=100, cluster_method="cluster_walktrap", propr=TRUE, sparcc=TRUE, cclasso=TRUE)
}
