% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phfunc.R
\name{phfunc}
\alias{phfunc}
\title{phfunc}
\usage{
phfunc(
  otu,
  taxa,
  sample,
  tree,
  functional_profile,
  pre_threshold,
  hcol = "red",
  ttip = 3,
  tlab = 2,
  hlab = 4,
  llab = 10,
  alab = 4,
  offset = 0.3,
  width = 1,
  mini_abun,
  threshold,
  sample_name,
  sample_group,
  mini_tpm,
  threshold_tpm
)
}
\arguments{
\item{otu}{a community count data matrix with samples in rows and OTUs/taxa in column,}

\item{taxa}{a taxa information matrix. The rownames must match the OTU names (taxa_names) of the otu,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{tree}{a phylogenetic tree. The tip labels must match the otu names in the otu,}

\item{functional_profile}{a community count data matrix with samples in rows and OTUs/taxa in column,}

\item{pre_threshold}{occurrence frequency,}

\item{hcol}{the color of heatmap,}

\item{ttip}{the node size of the tree,}

\item{tlab}{the font size of the tree tip label,}

\item{hlab}{the font size of the heatmap label,}

\item{llab}{the font size of the legend label,}

\item{alab}{the font size of the title,}

\item{offset}{the offset between each subfigure,}

\item{width}{the width of each heatmap cell,}

\item{mini_abun}{a value indicating whether the otu is present,}

\item{threshold}{a value indicating the final percent increase in beta-diversity,}

\item{sample_name}{the name of a column that containing the sample name,}

\item{sample_group}{the name of a column that containing the sample type/group,}

\item{mini_tpm}{a value indicating whether the function is present,}

\item{threshold_tpm}{a value indicating the final percent increase in beta-diversity,}
}
\value{
Combined plot showing the phylogenetic tree and the heatmaps.
}
\description{
Constructs a plot combining the phylogenetic tree, the heatmaps showing average occurrence frequency and average relative abundance of each taxa.
}
\details{
Constructs a plot combining the phylogenetic tree, the heatmaps showing average occurrence frequency and average relative abundance of each taxa.
}
\examples{
 phfunc(otu, taxa, sample, tree, functional_profile, pre_threshold, hcol = 'red', ttip = 3, tlab = 2, hlab = 4, llab = 4, offset = 0.3, width = 1, mini_abun = 0, threshold = 0.02, sample_name, sample_group, mini_tpm = 1000, threshold_tpm = 0.5)
}
