% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phnetworks.R
\name{phnetworks}
\alias{phnetworks}
\title{phnetworks}
\usage{
phnetworks(
  otu,
  taxa,
  sample,
  tree,
  propr = TRUE,
  sparcc = TRUE,
  cclasso = TRUE,
  pre_threshold,
  fdr_threshold,
  cor_threshold,
  permutation,
  nscore = "degree",
  hcol = "red",
  propr_col = "darkcyan",
  sparcc_col = "firebrick3",
  cclasso_col = "goldenrod3",
  ttip = 3,
  tlab = 2,
  hlab = 4,
  llab = 4,
  offset = 0.3,
  width = 1,
  mini_abun,
  threshold,
  sample_name,
  sample_group
)
}
\arguments{
\item{otu}{a community count data matrix,}

\item{taxa}{a taxa information matrix. The rownames must match the OTU names (taxa_names) of the otu,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{tree}{a phylogenetic tree. The tip labels must match the otu names in the otu,}

\item{propr}{a logical value indicating whether to run propr,}

\item{sparcc}{a logical value indicating whether to run sparcc,}

\item{cclasso}{a logical value indicating whether to run cclasso,}

\item{pre_threshold}{occurrence frequency,}

\item{fdr_threshold}{fdr threshold,}

\item{cor_threshold}{association threshold,}

\item{permutation}{permutation number of the association methods,}

\item{nscore}{the node-level property (degree, betweenness and closeness),}

\item{hcol}{the color of heatmap,}

\item{propr_col}{the color of the properties of propr network,}

\item{sparcc_col}{the color of the properties of sparcc network,}

\item{cclasso_col}{the color of the properties of cclasso network,}

\item{ttip}{the node size of the tree,}

\item{tlab}{the font size of the tree tip label,}

\item{hlab}{the font size of the heatmap label,}

\item{llab}{the font size of the legend label,}

\item{offset}{the offset between each subfigure,}

\item{width}{the width of each heatmap cell,}

\item{mini_abun}{a value indicating whether the otu is present,}

\item{threshold}{a value indicating the final percent increase in beta-diversity,}

\item{sample_name}{the name of a column that containing the sample name,}

\item{sample_group}{the name of a column that containing the sample type/group,}
}
\value{
Combined plot showing the phylogenetic tree, the heatmap and co-occurrence networks.
}
\description{
Constructs a plot combining the phylogenetic tree, the heatmaps showing average occurrence frequency and average relative abundance of each taxa, and the co-occurrence networks from different methods.
}
\details{
Constructs a plot combining the phylogenetic tree, the heatmaps showing average occurrence frequency and average relative abundance of each taxa, and the co-occurrence networks from different methods.
}
\examples{
 phnetworks(otu, taxa, sample, tree, propr = TRUE, sparcc = TRUE, cclasso = TRUE, pre_threshold, fdr_threshold, cor_threshold, permutation, nscore = 'degree', hcol = 'red', propr_col = 'darkcyan', sparcc_col = 'firebrick3', cclasso_col = 'goldenrod3', ttip = 3, tlab = 2, hlab = 4, llab = 4, offset = 0.3, width = 1, mini_abun = 0, threshold = 0.02, sample_name, sample_group)
}
