% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preve_func_dis_plot.R
\name{preve_func_dis_plot}
\alias{preve_func_dis_plot}
\title{preve_func_dis_plot}
\usage{
preve_func_dis_plot(
  functional_profile,
  sample,
  mini_tpm,
  threshold,
  core_col,
  noncore_col
)
}
\arguments{
\item{functional_profile}{a community count data matrix with samples in rows and OTUs/taxa in column,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{mini_tpm}{a value indicating whether the function is present,}

\item{threshold}{a value indicating the final percent increase in beta-diversity,}

\item{core_col}{the color of the core,}

\item{noncore_col}{the color of the non-core,}
}
\value{
Two density plots showing the average occupancy and average TPM between core and non-core functional otu.
}
\description{
Plots average occupancy and average TPM between core and non-core functional otu.
}
\details{
Plots average occupancy and average TPM between core and non-core functional otu.
}
\examples{
 preve_func_dis_plot(functional_profile, sample, mini_abun=1000, threshold=0.5, core_col = "deeppink4", noncore_col = "darkseagreen3")
}
