% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{pval.sparccboot}
\alias{pval.sparccboot}
\title{This function get empirical p-values from bootstrap SparCC output
adapted from https://github.com/zdk123/SpiecEasi/blob/master/R/spaRcc.R}
\usage{
pval.sparccboot(x, sided = "both")
}
\description{
This function get empirical p-values from bootstrap SparCC output
adapted from https://github.com/zdk123/SpiecEasi/blob/master/R/spaRcc.R
}
