% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_plot.R
\name{random_forest_plot}
\alias{random_forest_plot}
\title{random_forest_plot}
\usage{
random_forest_plot(
  otu,
  sample,
  mini_abun,
  threshold,
  sample_name,
  sample_group
)
}
\arguments{
\item{otu}{a community count data matrix,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{mini_abun}{a value indicating whether the otu is present,}

\item{threshold}{a value indicating the final percent increase in beta-diversity,}

\item{sample_name}{the name of a column that containing the sample name,}

\item{sample_group}{the name of a column that containing the sample type/group,}
}
\value{
A bar plot showing importance of otu on predicting/classifying the sample_group and a bar plot showing the accuracy using core otu, non-core otu and all otu.
}
\description{
Plots the importance of otu on predicting/classifying the sample_group and the accuracy using core otu, non-core otu and all otu.
}
\details{
Plots the importance of otu on predicting/classifying the sample_group and the accuracy using core otu, non-core otu and all otu.
}
\examples{
 random_forest_plot(otu, sample, mini_abun=0, threshold=0.02, height=400, weight=400)
}
