% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{sparcc}
\alias{sparcc}
\title{This function runs SparCC in R
adapted from https://github.com/zdk123/SpiecEasi/blob/master/R/spaRcc.R}
\usage{
sparcc(data, iter = 20, inner_iter = 10, th = 0.1)
}
\description{
This function runs SparCC in R
adapted from https://github.com/zdk123/SpiecEasi/blob/master/R/spaRcc.R
}
