% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{sparccboot}
\alias{sparccboot}
\title{This function gets bootstrapped estimates of SparCC correlation coefficients
adapted from https://github.com/zdk123/SpiecEasi/blob/master/R/spaRcc.R}
\usage{
sparccboot(
  data,
  sparcc.params = list(),
  statisticboot = function(data, indices) triu(do.call("sparcc", c(list(data[indices, ,
    drop = FALSE]), sparcc.params))$Cor),
  statisticperm = function(data, indices) triu(do.call("sparcc",
    c(list(apply(data[indices, ], 2, sample)), sparcc.params))$Cor),
  R,
  ncpus = 1,
  ...
)
}
\description{
This function gets bootstrapped estimates of SparCC correlation coefficients
adapted from https://github.com/zdk123/SpiecEasi/blob/master/R/spaRcc.R
}
