% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_accumulation_plot.R
\name{species_acc_plot}
\alias{species_acc_plot}
\title{species_acc_plot}
\usage{
species_acc_plot(OTU_input)
}
\arguments{
\item{otu}{a community count data matrix,}
}
\value{
A species accumulation curve with boxplot indicating the 95\% CI.
}
\description{
Plots species accumulation curve with specaccum function in vegan package
}
\details{
Plots species accumulation curve with specaccum function in vegan package.
}
\examples{
 species_acc_plot(otu)
}
