% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_abun_plot.R
\name{tree_abun_plot}
\alias{tree_abun_plot}
\title{tree_abun_plot}
\usage{
tree_abun_plot(
  otu,
  sample,
  tree,
  mini_abun,
  threshold,
  sample_name,
  sample_group
)
}
\arguments{
\item{otu}{a community count data matrix,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{tree}{a phylogenetic tree. The tip labels must match the otu names in the otu,}

\item{mini_abun}{a value indicating whether the otu is present,}

\item{threshold}{a value indicating the final percent increase in beta-diversity,}

\item{sample_name}{the name of a column that containing the sample name,}

\item{sample_group}{the name of a column that containing the sample type/group,}
}
\value{
A phylogenetic tree with a heatmap denoting the average occurrence frequency of each taxa with color denoting and core or non-core otu.
}
\description{
Plots phylogenetic tree with a heatmap denoting the average occurrence frequency of each taxa.
}
\details{
Plots phylogenetic tree with a heatmap denoting the average occurrence frequency of each taxa.
}
\examples{
 tree_abun_plot(otu, sample, tree, mini_abun=0, threshold=0.02)
}
