% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_func_plot.R
\name{tree_func_plot}
\alias{tree_func_plot}
\title{tree_func_plot}
\usage{
tree_func_plot(
  functional_profile,
  sample,
  tree,
  mini_tpm,
  threshold,
  offset,
  width,
  core_col,
  noncore_col,
  low_col,
  high_col,
  hlab,
  ttip,
  tlab
)
}
\arguments{
\item{functional_profile}{a community count data matrix with samples in rows and OTUs/taxa in column,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{tree}{a phylogenetic tree. The tip labels must match the otu names in the otu,}

\item{mini_tpm}{a value indicating whether the function is present,}

\item{threshold}{a value indicating the final percent increase in beta-diversity,}

\item{offset}{the offset between each subfigure,}

\item{width}{the width of each heatmap cell,}

\item{core_col}{the color of core tips,}

\item{noncore_col}{the color of non-core tips,}

\item{low_col}{the lowest color of the heatmap,}

\item{high_col}{the highest color of the heatmap,}

\item{hlab}{the font size of the heatmap label,}

\item{ttip}{the node size of the tree,}

\item{tlab}{the font size of the tree tip label,}
}
\value{
A phylogenetic tree with a heatmap denoting the average occurrence frequency and average tpm of the functions of each taxa with color denoting and core or non-core functional otu.
}
\description{
Plots phylogenetic tree with a heatmap denoting the average occurrence frequency and average tpm of the functions of each taxa.
}
\details{
Plots phylogenetic tree with a heatmap denoting the average occurrence frequency and average tpm of the functions of each taxa.
}
\examples{
 tree_func_plot(functional_profile, sample, tree, mini_tpm=1000, threshold=0.5, offset = 0.1, width = 0.5, core_col = 'deeppink4', noncore_col = 'darkseagreen3', low_col = 'white', high_col = 'lightcoral', hlab = 2, ttip = 2, tlab = 2)
}
