% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_plot.R
\name{networks_plot}
\alias{networks_plot}
\title{networks_plot}
\usage{
networks_plot(
  otu,
  taxa,
  sample,
  pre_threshold,
  fdr_threshold,
  cor_threshold,
  permutation,
  propr,
  sparcc,
  cclasso,
  spieceasi,
  propr_col,
  sparcc_col,
  cclasso_col,
  spieceasi_col
)
}
\arguments{
\item{otu}{a community count data matrix with samples in rows and OTUs/taxa in column,}

\item{taxa}{a taxa information matrix. The rownames must match the OTU names (taxa_names) of the otu,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{pre_threshold}{occurrence frequency,}

\item{fdr_threshold}{fdr threshold,}

\item{cor_threshold}{association threshold,}

\item{permutation}{permutation number of the association methods,}

\item{propr}{a logical value indicating whether to run propr,}

\item{sparcc}{a logical value indicating whether to run sparcc,}

\item{cclasso}{a logical value indicating whether to run cclasso,}

\item{spieceasi}{a logical value indicating whether to run spieceasi,}

\item{propr_col}{the color of propr network,}

\item{sparcc_col}{the color of sparcc network,}

\item{cclasso_col}{the color of cclasso network,}

\item{spieceasi_col}{the color of spieceasi network,}
}
\value{
Co-occurrence network plots showing the relationship between otu.
}
\description{
Constructs and plots co-occurrence networks based on specific occurrence frequency (pre_threshold), permutation number (permutation), fdr threshold (fdr_threshold), association threshold (cor_threshold) with propr, sparcc and cclasso.
}
\details{
Constructs and plots networks based on specific occurrence frequency (pre_threshold), permutation number (permutation), fdr threshold (fdr_threshold), association threshold (cor_threshold) with propr, sparcc and cclasso.
}
\examples{
 networks_plot(otu, taxa, sample, pre_threshold=0.5, fdr_threshold=0.1, cor_threshold=0.6, permutation=100, propr=TRUE, sparcc=TRUE, cclasso=TRUE, spieceasi=TRUE, propr_col='darkcyan', sparcc_col='firebrick3', cclasso_col='goldenrod3', spieceasi_col = 'dodgerblue3')
}
