% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preve_abun_dis_plot.R
\name{preve_abun_dis_plot}
\alias{preve_abun_dis_plot}
\title{preve_abun_dis_plot}
\usage{
preve_abun_dis_plot(
  otu,
  sample,
  mini_abun,
  threshold,
  sample_name,
  sample_group,
  core_col,
  noncore_col
)
}
\arguments{
\item{otu}{a community count data matrix with samples in rows and OTUs/taxa in column,}

\item{sample}{a sample information data.frame. The rownames must match the sample names in the otu,}

\item{mini_abun}{a value indicating whether the otu is present,}

\item{threshold}{a value indicating the final percent increase in beta-diversity,}

\item{sample_name}{the name of a column that containing the sample name,}

\item{sample_group}{the name of a column that containing the sample type/group,}

\item{core_col}{the color of the core,}

\item{noncore_col}{the color of the non-core,}
}
\value{
Two density plots showing the average occupancy and average relative abundance between core and non-core otu.
}
\description{
Plots average occupancy and average relative abundance between core and non-core otu.
}
\details{
Plots average occupancy and average relative abundance between core and non-core otu.
}
\examples{
 preve_abun_dis_plot(otu, sample, mini_abun=0, threshold=0.02, sample_name, sample_group, core_col = "deeppink4", noncore_col = "darkseagreen3")
}
